Rewrite scheduling of operations
Before this change each AsyncFd has it's own allocation which could hold
the state for a single operation, mainly a place for the result and a
task::Waker. However this has a number of problems.

First, it means that an fd can only perform a single operation at a
time, even though io_uring can do multiple. An easy solution to this
seems to change the single allocated state to an array and move on. But
this is not as easy as it seems as we either have to decide how many
operations an fd can do concurrently up front (possible via some awkward
API) or dynamically grow. Dynamically growing is the nicer option of the
two, but that became very messy very quickly due to the user of raw
pointer to `Arc` allocated data passed to the kernel. It means that we
couldn't really reallocate the operation states during an operation.
After trying this, it just didn't work well.

The second problem is that cancelation and timeout need an additional
operation (state), blocking that work entirely.

Finally we make an allocation for each fd. Although allocations aren't
generally bad, for short lived operation this allocation isn't needed
(as shown by the new design).

The new design doesn't use a state per fd, instead it uses a single list
of queued operation states shared by all fds. This means that we don't
have an allocation per fd, just a larger one made up front. Furthermore
it means we can have multiple operations per fd, unblocking future work
such as timeouts and cancelation. Finally it reuses (would be)
allocations we an fd doesn't have an active operation.

The list of operation state lives inside of SubmissionQueue, the
queued_ops field. To quickly determine an empty slot into the queue we
use the AtomicBitMap type added in a previous commit, stored as
op_indices. To schedule an operation we now check op_indices for a slot
into queued_ops to store the operation state into.