Add test_sanitizers target to Makefile
This will return the tests with all support sanitizers.

For the address and therads sanitizers it currently find no issues. For
the leak sanitizer however I ran into an issue, I'm not able to quickly
resolve and fancy spending more time on because it's also part of the
memory sanitizer. However the memory sanitizer also doesn't work because
it doesn't understand io_uring and claims we're reading a bunch of
uninitialised memory. So, both the leak and memory sanitizer are
currently disabled.