Improve performance of AtomicBitMap::next_available
The previous code would check one single bit in one iteration. This
changes that to use trailing_ones to skip all the bits set to one and
directly do the atomic fetch_or. This removes all iterations that would
fail the bit unset check, reducing the number of iterations drasticly
for densly populated bitmaps.

This also adds a new test that does setting and unsetting of the bits
concurrently.