Remove QueueFull error from API
Instead handle it internall in the Future. This has two major changes.

First, the API of all Future creating methods and function changed to no
longer return a result, but the Future implementation directly.
Futhermore the QueueFull becomes private.

Second, the Futures are now lazy, meaning they are only started when the
Future is polled. This has some side effects such as
AsyncFd::cancel_previous working differently when the to-be-canceled
Future wasn't polled yet vs. when it's polled (and the operation is
queued). This behaviour is now documented in the
AsyncFd::cancel_previous method.

Also merges the SyncAll and SyncData Future types, as they are more or
less the same (only the flags are different), matching AsyncFd::send and
AsyncFd::send_zc.

To support this setup, wherein the Futures do the setup, not the method
creating them, we need a way wake the Futures that fail add a submission
due to QueueFull errors. Because I except most non-example/demo Rings to
be fairly large, at least large enough to handle all concurrent Futures,
the implementation isn't performance focused. It's a simple Vector
behind a lock.

When a Future fails to add a submission it adds itself to the this
vector, for now (poorly) called `blocked_futures` on the
SharedSubmissionQueue. When Ring::poll is called this vector is clearly
and all wakers are awoken.

This is a bad implementation not only due the use of the Mutex, but more
so due to the waking strategy. The Ring::poll assumes that if the
completions are empties it also free slots in the sumbission queue, but
this doesn't have to true. It makes we could be waking Future that can
make no progress because the submission queue is (still) full.

Finally this makes some small edits to op_future! outside of those
required to support this new implementation. Mainly it adds
`map_result: ` before the mapping function to indicate it's use as the
macro call is getting a little busy.