Support multishot operations in QueuedOperation
This removes some complexity from the type and Ring::poll and the
functions it has to call to process a completion event. Instead of
partially handling the event in Ring::poll and QueuedOperation's method
we now do it all in a single function: QueuedOperation::update. This
removes the OperationResult enum entirely and replaces it with
QueuedOperationKind and some boolean fields on QueuedOperation.

The QueuedOperationKind enum is required to support multishot
operations, which can result in multiple operation results before the
user has a chance to read them.

The downside of support multishot operations is the growth in the size
of QueuedOperation from 24 to 56 bytes. Most of this comes from the
added Vec (24 bytes), the QueuedOperationKind enum tag (8 bytes) and
the additional booleans (8 bytes).

We could reduce the size QueuedOperationKind using something like a
ThinVec which holds its capacity and length in the heap allocation
instead of the stack. Something to consider in the future.