Don't use async_iter feature
This adds a "nightly" feature which allows us to depend on nightly-only
features such as async_iter.

When this feature is not enable we no longer use the AsyncIterator trait
and instead expose a public poll_next function which holds the same
implementation and function signature. This allows A10 to be used on
stable Rust, but still support the AsyncIterator trait when the nightly
feature is enabled.