Make cancel_all tests more robust
Futures are inert and we can't determine if an operation has actually
been started or the submission queue was full. This means that we can't
ensure that the operation has been queued with the Kernel. This means
that in some cases we won't actually cancel any operations simply
because they haven't started. This introduces flakyness in the tests.

To work around this adds a cancel_all function. If we fail to get the
expected number of canceled operations we try to start the operations
again, before canceling them again. Looping until we get the expected
number of canceled operations.