Use DropWaker SubmissionQueue::cancel_op
Instead of always using the Box implementation, meaning we always
allocate, we now require the caller to pass resources that implement
DropWaker.

This means we can avoid the allocation for various types that implement
DropWaker already.

The downside is that in the case where we got no more events the
allocation is made upfront (for certain types) and then dropped, without
actually being in a waker.