Move everything in a io_uring module
Breaks everything, but now it's all moved. From this point we can start
building up a cross-platform API in root of the crate.

The tests are not updated as we want roughly the API as the old code, so
we don't want to change them only to change them back.