Rewrite sq::Queue::submit
Now requires a task::Waker, which is needed to create QueuedOperation
(as that was made quired in the previous commit).

Splits the submitting to the completion queue implementation to a new
function: submit_with_id. This will be used in a new function as well.

submit_with_id will automatically use the queued operation's waker to
wait for a submission.