Change logging for unexpected completion events
For wake ups we can simple ignore the completion event.

For submission for which we didn't expect a completion event it means
they failed. However since we can't related them back to a specific
operation all we can do is log them.

For all other we keep using the same trace log.