Add fd_iter_operation! macro
Implements AsyncIterator using FdOperation.

Requires a new trait, FdIter, which extends FdOp with map_next, which is
similar to FdOp::map_ok, but accepts the resources a mutable reference
as the resources are needed in the next calls.

Also adds State::poll_next to support FdOperation's poll_next function.
It's similar to State::poll, but implements an AsyncIterator instead of
a Future.

Also extends the new_operation! macro to implement the AsyncIterator
trait for the created type.