Properly cancel operations
Introduces a new operation state: State::Cancelled which is set if the
operation has not been started and then cancel is called on it.

Before this would effectively do nothing as the operation was not
started, thus it also woudln't be cancelled. But polling the
future/async iter afterwards would let it start as if nothing happened.