Read completion queue head using Acquire ordering
We, as in the Completions type, should be the only one writing to it.
Furthermore the Completions type is uniquely owned by the Ring type (via
cq::Queue). All of this should point to relaxed ordering being fine, but
at this point I'm rather safe than sorry and change this to acquire.