Fix io_uring::Shared::unsubmitted
We need to read the head (written by the kernel) before the tail
(written by us) to ensure it's not possible that the head is larger than
the tail.

This is possible if
1. we read the tail first,
2. another threads submits an operation,
3. the kernel processes the submission and updates the head
4. we read the head

We've read tail = N, while it's updated to N + 1, and we read the
head = N + 1. Which means that tail - head = -1.