Reduce size of point in SocketAddr::init
Since the underlying storage can either be sockaddr_in or sockaddr_in6,
if we use a pointer to sockaddr_in6 it's possible it's not fully
initialised. Instead used a pointer to sa_family_t, which is the first
field of sockaddr_in and sockaddr_in6 and all we need at this point in
the code.