Reduce the time we indicate we're polling
This makes two changes. First, it moves the setting of is_polling to the
implementation specific code. This already reduces the time we set
polling. Especially for the kqueue implementation that check all events
for errors.

The second change is that we don't set is_polling at all if the timeout
is zero, i.e. we're not waiting. For the io_uring implementation
specifically the reduces the time that is_polling is set to true.

Combined this should reduce the amount of of wakes that are done.