Add fs::notify
This adds a new fs::notify module that implements watching of file
system changes using inotify.

The main type is fs::notify::Watcher, which does the actual watching of
the file system. Adding files and directories, using watch_file and
watch_directory, is done using synchronous I/O to make it easier to
setup intially. Later it should be possible to add function to
asynchronously add new watches.

The read of events is done asynchronously and implemented as
AsyncIterator (see Events).