Switch watch order in Watcher::watch_directory
When using the recusrive watching of a directory we used to first watch
the directory itself, then traverse it and watch all it's
sub directories.

However, this would cause a bunch of events to be triggered for opening
the directory and sub files. Basically, we created a bunch of false
positives while setting up the watcher.

This commit changes the watching to first traverse the directory and
watch all it's sub-directories and files and only then start watching
the directory itself. This way no events are generated when adding
recursive watches.